requirements.txt
google-genai>=0.1.0
Pillow>=10.0.0
rawpy>=0.18.0
numpy>=1.21.0
ShopifyAPI>=12.0.0
requests>=2.28.0
pandas>=1.5.0
gdown>=4.7.0

config~=0.5.1